Created GameActivity to use fragments for the game. Starting VIPER structure
