+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
           package="com.pedrodavidlp.ittrivial">
+
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
@@ -8,26 +9,29 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".login.view.MenuActivity">
-        </activity>
-        <activity android:name=".game.view.activity.GameActivity">
-        </activity>
-        <activity android:name=".login.view.EnterGameActivity">
+        <fragment android:name=".login.view.MenuActivity">
+        </fragment>
+        <fragment android:name=".game.view.fragment.RouletteFragment">
+        </fragment>
+        <fragment android:name=".login.view.EnterGameActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
 
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
-        </activity>
-        <activity android:name=".game.view.activity.QuestionActivity">
-        </activity>
-        <activity android:name=".login.view.PlayerListAdminActivity">
-        </activity>
-        <activity android:name=".login.view.PlayerListGuestActivity">
-        </activity>
-        <activity android:name=".login.view.GameListActivity">
-        </activity>
-        <activity android:name=".game.view.activity.WaitActivity">
+        </fragment>
+        <fragment android:name=".game.view.fragment.QuestionFragment">
+        </fragment>
+        <fragment android:name=".login.view.PlayerListAdminActivity">
+        </fragment>
+        <fragment android:name=".login.view.PlayerListGuestActivity">
+        </fragment>
+        <fragment android:name=".login.view.GameListActivity">
+        </fragment>
+        <fragment android:name=".game.view.fragment.WaitActivity">
+        </fragment>
+
+        <activity android:name=".game.view.activity.GameActivity">
         </activity>
     </application>
