+++ app/src/main/java/com/pedrodavidlp/ittrivial/game/view/activity/transition/TransitionSoftware.kt
@@ -2,7 +2,7 @@ package com.pedrodavidlp.ittrivial.game.view.activity.transition
 
 import com.pedrodavidlp.ittrivial.game.view.Category
 
-class TransitionSoftware(image: android.widget.ImageView, router: com.pedrodavidlp.ittrivial.game.router.GameRouter) : TransitionTemplate(image, router) {
+class TransitionSoftware(image: android.widget.ImageView, router: com.pedrodavidlp.ittrivial.game.router.RouletteRouter) : TransitionTemplate(image, router) {
   override fun goToQuestion() {
     router.goToQuestion(Category.SOFTWARE)
   }
