+++ app/src/main/java/com/pedrodavidlp/ittrivial/game/view/activity/transition/TransitionTemplate.kt
@@ -5,9 +5,9 @@ import android.view.View.VISIBLE
 import android.view.animation.AlphaAnimation
 import android.view.animation.Animation
 import android.widget.ImageView
-import com.pedrodavidlp.ittrivial.game.router.GameRouter
+import com.pedrodavidlp.ittrivial.game.router.RouletteRouter
 
-abstract class TransitionTemplate(val image: ImageView, val router: GameRouter) {
+abstract class TransitionTemplate(val image: ImageView, val router: RouletteRouter) {
 
   fun makeTransition() {
     setImage()
