+++ app/src/main/res/layout/activity_question.xml
@@ -5,7 +5,7 @@
                                              android:id="@+id/container"
                                              android:layout_width="match_parent"
                                              android:layout_height="match_parent"
-                                             tools:context="com.pedrodavidlp.ittrivial.game.view.activity.QuestionActivity">
+                                             tools:context="com.pedrodavidlp.ittrivial.game.view.fragment.QuestionFragment">
 
     <Button
         android:id="@+id/first"
