+++ app/src/main/res/layout/activity_wait.xml
@@ -4,7 +4,7 @@
                                              xmlns:tools="http://schemas.android.com/tools"
                                              android:layout_width="match_parent"
                                              android:layout_height="match_parent"
-                                             tools:context="com.pedrodavidlp.ittrivial.game.view.activity.WaitActivity">
+                                             tools:context="com.pedrodavidlp.ittrivial.game.view.fragment.WaitActivity">
 
     <TextView
         android:id="@+id/waitYourTurnText"
